requirements/test_requirements.txt
django-test-plus==1.0.14
factory-boy==2.7.0
fake-factory==0.5.7
ipaddress==1.0.16
python-dateutil==2.5.3
freezegun==0.3.7
